requirements.txt
streamlit
transformers
torch
pdfplumber
python-docx